requirements.txt
Jinja2
msgpack-python > 0.3
PyYAML
MarkupSafe
requests >= 1.0.0
PyHamcrest >=1.8.2